+++ crossport/src/main/java/com/webapplication/crossport/ui/controllers/CategoryController.java
@@ -42,28 +42,32 @@ public class CategoryController {
     }
 
     @PostMapping("")
-    public String add(final @Valid CategoryDTO categoryDTO, final BindingResult bindingResult, final Model model) {
-        Category sameCategory = categoryService.getFirstByName(categoryDTO.getCategoryName());
-
-        if (sameCategory != null) {
-            bindingResult.addError(new ObjectError("globalError", "Same category already exists."));
-        }
-
-        if (bindingResult.hasErrors()) {
-            model.addAttribute("categoryDTO", categoryDTO);
-        } else {
-            Category category = new Category();
-            category.setName(categoryDTO.getCategoryName());
-            categoryService.saveCategory(category);
+    public String add(final @Valid CategoryDTO categoryDTO,
+                      final BindingResult bindingResult,
+                      final Model model) {
+        try {
+            categoryService.addCategory(categoryDTO);
+        } catch (RuntimeException e) {
+            if (!bindingResult.hasErrors()) {
+                bindingResult.addError(new ObjectError("Error", e.getMessage()));
+            }
         }
 
         model.addAttribute("listCategories", categoryService.getAllCategories());
+        model.addAttribute("categoryDTO", categoryDTO);
         return "categories";
     }
 
     @GetMapping("/{id}")
     public String getById(@PathVariable(value = "id") Integer id, Model model) {
-        Category category = categoryService.getCategoryById(id);
+        Category category;
+
+        try {
+            category = categoryService.getCategoryById(id);
+        } catch (RuntimeException e) {
+            return "categories";
+        }
+
         List<Article> articlesNotInCategory = articleService.getArticlesNotInCategory(category);
 
         model.addAttribute("category", category);
@@ -75,24 +79,13 @@ public class CategoryController {
     public String delete(@PathVariable(value = "id") Integer id,
                          @RequestParam(value = "confirm", defaultValue = "false") boolean confirm,
                          RedirectAttributes redir) {
-        Category category;
         try {
-            category = categoryService.getCategoryById(id);
+            categoryService.deleteCategory(id, confirm);
         } catch (RuntimeException e) {
-            return "redirect:/categories";
-        }
-
-        if (category.getArticles().isEmpty() || confirm) {
-            try {
-                categoryService.deleteCategory(id);
-            } catch (RuntimeException e) {
-                return "redirect:/categories";
-            }
-        } else {
-            String delError = "You cannot delete this category as it has articles bound.";
-            redir.addFlashAttribute("delError", delError);
+            redir.addFlashAttribute("delError", e.getMessage());
             return "redirect:/categories/" + id;
         }
+
         redir.addFlashAttribute("listCategories", categoryService.getAllCategories());
         redir.addFlashAttribute("categoryDTO", new CategoryDTO());
         return "redirect:/categories";
@@ -102,21 +95,12 @@ public class CategoryController {
     public String removeArticle(@PathVariable(value = "idCategory") Integer idCategory,
                                 @PathVariable(value = "idArticle") Integer idArticle,
                                 RedirectAttributes redir) {
-        Category category;
         try {
-            category = categoryService.getCategoryById(idCategory);
+            articleService.removeCategory(idArticle, idCategory);
         } catch (RuntimeException e) {
-            String delError = "An error occured when we tried to find the category selected";
-            redir.addFlashAttribute("delError", delError);
-            return "redirect:/categories/" + idCategory;
-        }
-
-        try {
-            articleService.removeCategory(idArticle, category);
-        } catch (RuntimeException e) {
-            String delError = "An error occured when we tried to delete the category from the article selected";
-            redir.addFlashAttribute("delError", delError);
-            return "redirect:/categories/" + idCategory;
+            redir.addFlashAttribute("delError",
+                    "An error occured when we tried to " +
+                            "delete the category from the article selected");
         }
 
         return "redirect:/categories/" + idCategory;
@@ -126,21 +110,12 @@ public class CategoryController {
     public String addArticle(@PathVariable(value = "idCategory") Integer idCategory,
                              @PathVariable(value = "idArticle") Integer idArticle,
                              RedirectAttributes redir) {
-        Category category;
         try {
-            category = categoryService.getCategoryById(idCategory);
+            articleService.addCategory(idArticle, idCategory);
         } catch (RuntimeException e) {
-            String delError = "An error occured when we tried to find the category selected";
-            redir.addFlashAttribute("delError", delError);
-            return "redirect:/categories/" + idCategory;
-        }
-
-        try {
-            articleService.addCategory(idArticle, category);
-        } catch (RuntimeException e) {
-            String delError = "An error occured when we tried to add the category from the article selected";
-            redir.addFlashAttribute("delError", delError);
-            return "redirect:/categories/" + idCategory;
+            redir.addFlashAttribute("delError",
+                    "An error occured when we tried to " +
+                            "add the category from the article selected");
         }
 
         return "redirect:/categories/" + idCategory;
